rhn unit tests: use mock_request as a fixture
