--- test/units/modules/packaging/utils.py
@@ -1,36 +1,0 @@
-import contextlib
-import json
-
-from ansible.compat.tests.mock import patch
-from ansible.module_utils import basic
-from ansible.module_utils.six.moves import xmlrpc_client
-from ansible.module_utils._text import to_bytes
-
-import pytest
-
-
-def get_method_name(request_body):
-    return xmlrpc_client.loads(request_body)[1]
-
-
-@pytest.fixture
-def mock_request(mocker, responses, module_name):
-    @contextlib.contextmanager
-    def _mock_request():
-        def transport_request(host, handler, request_body, verbose=0):
-            """Fake request"""
-            method_name = get_method_name(request_body)
-            excepted_name, response = responses.pop(0)
-            if method_name == excepted_name:
-                if isinstance(response, Exception):
-                    raise response
-                else:
-                    return response
-            else:
-                raise Exception('Expected call: %r, called with: %r' % (excepted_name, method_name))
-
-        target = '{0}.xmlrpc_client.Transport.request'.format(module_name)
-        with mocker.mock_module.patch(target, side_effect=transport_request) as e:
-            yield e
-
-    return _mock_request
